Move rust generation module
