+++ crates/formality-rust/src/grammar.rs
@@ -37,5 +37,3 @@ mod wc;
 pub use wc::*;
 
 pub mod expr;
-
-pub mod rust_builder;
