+++ crates/formality-rust/src/lib.rs
@@ -48,6 +48,7 @@ pub mod grammar;
 pub mod prove;
 mod test;
 pub mod test_util;
+pub mod to_rust;
 mod trait_binder;
 pub mod types;
